+++ routes/imageViewerRoutes.js
@@ -65,6 +65,7 @@ router.post("/images/:id/rating", imageViewerController.updateRating);
 router.post("/images/:id/ranking", imageViewerController.updateRanking);
 router.post("/images/:id/review", imageViewerController.updateReview);
 router.delete("/images/:id", imageViewerController.deleteImage);
+router.get("/subfolders", imageViewerController.getSubfolders);
 router.post("/scan", imageViewerController.scanImages);
 router.get("/tags", imageViewerController.getTags);
 router.get("/info", imageViewerController.getInfo);